Replace separator in input
It is possible that the separator appears in the input data, to be
printed to a file. This was not filtered out and thus the CSV output
could sometimes have too many columns.
Replacing the separator with another character is not the best solution.
Normally you want to put the string containing the separator in quotes,
so that parsers understand that the whole string is for one column.