Remove unmatched quotes
With a dirty quote `"` that is not matched, most CSV parsers wil
interpret this as the start of a literal. All the following data will be
considered data for the same "cell". In this case there is not much more
we can do than just remove all the quotes. In other cases (matched
quotes), they can safely stay.